Moved actions to FlatTextFieldUI
